+++ .buildkite/scripts/steps/eslint.ts
@@ -28,7 +28,6 @@ void (async () => {
   // TODO: fix this to where we can install only the necessary packages in one script
   await yarnInstall('e2e');
   await yarnInstall('.buildkite');
-  await yarnInstall('github_bot');
 
   // TODO: fix this to avoid requiring build to run linting
   // currently the storybook and others are loosely coupled to @elastic/charts/src