+++ packages/charts/src/utils/themes/theme.ts
@@ -28,6 +28,7 @@ export interface TextStyle {
   fontSize: number;
   fontFamily: string;
   fontStyle?: FontStyle;
+  fontWeight?: FontWeight;
   fill: Color;
   padding: number | SimplePadding;
 }
@@ -305,7 +306,7 @@ export interface HeatmapStyle {
  * Metric font weight options for text styling.
  * @public
  */
-export type MetricFontWeight = Extract<FontWeight, 'bold' | 'normal'>;
+export type MetricFontWeight = Extract<FontWeight, 'bold' | 'normal' | 500>;
 
 /**
  * Style options for the Metric chart type.
@@ -402,7 +403,16 @@ export interface LegendLabelOptions {
    *
    * @defaultValue 1
    */
-  maxLines: number;
+  maxLines?: number;
+  /**
+   * Sets widthLimit allowable before truncating (unit: px)
+   * Only applicable if the list layout is chosen for the top and bottom positions
+   *
+   * Setting value to `0` will _never_ truncate the text
+   *
+   * @defaultValue 250
+   */
+  widthLimit?: number;
 }
 
 /** @public */